requirements.txt
openai
tiktoken
langchain
python-dotenv
requests
unstructured
markdown
qdrant-client
chromadb
azure-search-documents==11.4.0b6
azure-identity